webstorage/storage-key-empty-string.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>WebStorage Test: Storage - key with empty string</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/webstorage/#the-storage-interface">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  ["localStorage", "sessionStorage"].forEach(function(name) {
    test(function () {
      var storage = window[name];
      storage.clear();

      storage.setItem("", "empty string");
      assert_equals(storage.getItem(""), "empty string");

    }, name + ": Check if Storage.key with empty string is valid");
  });
</script>
